decArgo_20220217_047a commit

+++ decArgo_soft/soft/sub/get_prof_param_fill_value.m
@@ -26,16 +26,16 @@ o_paramFillValueList = [];
 if (isempty(a_profData.paramNumberWithSubLevels))
    for idParam = 1:length(a_profData.paramList)
       paramInfo = get_netcdf_param_attributes(a_profData.paramList(idParam).name);
-      o_paramFillValueList = [o_paramFillValueList paramInfo.fillValue];
+      o_paramFillValueList = [o_paramFillValueList double(paramInfo.fillValue)];
    end
 else
    for idParam = 1:length(a_profData.paramList)
       paramInfo = get_netcdf_param_attributes(a_profData.paramList(idParam).name);
       if (any(a_profData.paramNumberWithSubLevels == idParam))
          idF = find(a_profData.paramNumberWithSubLevels == idParam);
-         o_paramFillValueList = [o_paramFillValueList repmat(paramInfo.fillValue, 1, a_profData.paramNumberOfSubLevels(idF))];
+         o_paramFillValueList = [o_paramFillValueList repmat(double(paramInfo.fillValue), 1, a_profData.paramNumberOfSubLevels(idF))];
       else
-         o_paramFillValueList = [o_paramFillValueList paramInfo.fillValue];
+         o_paramFillValueList = [o_paramFillValueList double(paramInfo.fillValue)];
       end
    end
 end
